+++ src/pages/SearchPage.tsx
@@ -2,9 +2,12 @@ import Slide from '@/components/Slide';
 import Layout from '@/components/Layout';
 import { useMovieState } from '@/context/MovieContext';
 import MovieSearch from '@/components/MovieSearch';
+import NoData from '@/components/Bookmark/NoData';
 
 const SearchPage = () => {
   const { movies } = useMovieState();
+
+  if (!movies.length) return <NoData />;
 
   return (
     <Layout>
